feat: dynamic length event
Change from a fixed type for events from BPF to a variable length one.
This is done by using a helper map in which we can write all the
information for a given event, then submit it to the ringbuffer with the
bpf_ringbuf_output helper.

On the userspace, we get a RingBufItem which can be treated as a slice
of bytes we can split to retrieve fields one by one.

This change needs to be properly tested for performance, we should use a
lot less space for each event in the ringbuffer but serializing and
deserializing each event might require more CPU.

TODO: Document the format for the serialized event in the ringbuffer.
TODO: Add tests for the event::parser module.

+++ fact/src/host_scanner.rs
@@ -174,8 +174,8 @@ impl HostScanner {
 
         let metadata = path.metadata()?;
         let inode = inode_key_t {
-            inode: metadata.st_ino(),
-            dev: metadata.st_dev(),
+            inode: metadata.st_ino() as u32,
+            dev: metadata.st_dev() as u32,
         };
 
         let host_path = host_info::remove_host_mount(path);
